Implementing services in the application
In this commit, the IServiceProduct and IProduct interfaces were created
for the implementation of the ServiceProduct and Product classes. For
the best abstraction of the code, the config.php file was also created
for our configuration and service.php for the implementation of the
services. File list.product.php presenting the template of our
application, so you can leave the index cleaner.